requirements.txt
pillow
flask
tensorflow
numpy
pandas
opencv-python
matplotlib
scikit-learn
streamlit

